+++ crates/aranya-runtime/src/storage/memory.rs
@@ -326,10 +326,10 @@ impl Storage for MemStorage {
             FactPerspectivePrior::FactPerspective(prior) => Some(self.write_facts(*prior)?),
             FactPerspectivePrior::FactIndex(prior) => Some(prior),
         };
-        if facts.map.is_empty() {
-            if let Some(prior) = prior {
-                return Ok(prior);
-            }
+        if facts.map.is_empty()
+            && let Some(prior) = prior
+        {
+            return Ok(prior);
         }
         Ok(MemFactIndex(Arc::new(MemFactsInner {
             map: facts.map,
@@ -340,10 +340,10 @@ impl Storage for MemStorage {
     fn commit(&mut self, segment: Self::Segment) -> Result<(), StorageError> {
         // TODO(jdygert): ensure segment belongs to self?
 
-        if let Some(head) = self.head {
-            if !self.is_ancestor(head, &segment)? {
-                return Err(StorageError::HeadNotAncestor);
-            }
+        if let Some(head) = self.head
+            && !self.is_ancestor(head, &segment)?
+        {
+            return Err(StorageError::HeadNotAncestor);
         }
 
         self.head = Some(segment.head_location());